modificaciones

+++ pantallas/lista_categoria.php
@@ -61,16 +61,19 @@
                             Editar
                         </button>
                          </a>
+                         <?php if ($fila["estatus"] == 0): ?>
                           <a href="../controladores/eliminar_categoria.php?id=<?=$fila['id']?>">
                             <button type="submit" class="bg-red-600 hover:bg-red-700 text-white px-3 py-1 rounded">
                                 Eliminar
                             </button>
                         </a>
+                        <?php else: ?>
                         <a href="../controladores/activar_categoria.php?id=<?=$fila['id']?>">
                             <button type="submit" class="bg-blue-600 hover:bg-blue-700 text-white px-3 py-1 rounded">
                                 activar
                             </button>
                         </a>
+                        <?php endif; ?>
                     </td>
                 </tr>
